Change the location of keystore to My Documents for Desktop systems.
Majority of backup solutions don't include AppData or LocalLow due to the amount of files and cache directories. I'm making this change to help reduce people frustration when the file itself gets corrupt.

Includes a system to migrate from the old and into the new location. Still uses the normal CompanyName/ProductName.